refactor: split upgrade data
